+++ packages/atomic/cypress/e2e/facets/automatic-facet-generator/automatic-facet-generator-assertions.ts
@@ -1,0 +1,24 @@
+import {automaticFacetComponent} from '../automatic-facet/automatic-facet-selectors';
+
+export const automaticFacetGeneratorComponent =
+  'atomic-automatic-facet-generator';
+
+export function assertContainsAutomaticFacet() {
+  cy.get(automaticFacetGeneratorComponent).find(automaticFacetComponent);
+}
+
+export function assertCollapseAutomaticFacets(isCollapsed: boolean) {
+  cy.get(automaticFacetGeneratorComponent)
+    .find(automaticFacetComponent)
+    .shadow()
+    .find('[part="values"]')
+    .should(isCollapsed ? 'not.exist' : 'be.visible');
+}
+
+export function assertDisplayPlaceholder() {
+  cy.get(automaticFacetGeneratorComponent).find('[part="placeholder"]');
+}
+
+export function assertDisplayNothing() {
+  cy.get(automaticFacetGeneratorComponent).children().should('not.exist');
+}