+++ packages/atomic/cypress/e2e/facets/automatic-facet-generator/automatic-facet-generator.cypress.ts
@@ -3,22 +3,24 @@ import {
   assertConsoleError,
   assertContainsComponentError,
 } from '../../common-assertions';
-import {addAutomaticFacetBuilder} from './automatic-facet-builder-actions';
+import {addAutomaticFacetGenerator} from './automatic-facet-generator-actions';
 import {
   assertCollapseAutomaticFacets,
   assertContainsAutomaticFacet,
-  automaticFacetBuilderComponent,
-} from './automatic-facet-builder-assertions';
+  assertDisplayNothing,
+  assertDisplayPlaceholder,
+  automaticFacetGeneratorComponent,
+} from './automatic-facet-generator-assertions';
 
-describe('Automatic Facet Builder Test Suites', () => {
+describe('Automatic Facet Generator Test Suites', () => {
   it('should throw an error when desiredCount property is invalid', () => {
     new TestFixture()
-      .with(addAutomaticFacetBuilder({'desired-count': 'potato'}))
+      .with(addAutomaticFacetGenerator({'desired-count': 'potato'}))
       .init();
     assertConsoleError();
     assertContainsComponentError(
       {
-        shadow: () => cy.get(automaticFacetBuilderComponent).shadow(),
+        shadow: () => cy.get(automaticFacetGeneratorComponent).shadow(),
       },
       true
     );
@@ -26,19 +28,19 @@ describe('Automatic Facet Builder Test Suites', () => {
 
   it('should display atomic-automatic-facet when desiredCount is valid', () => {
     new TestFixture()
-      .with(addAutomaticFacetBuilder({'desired-count': '1'}))
+      .with(addAutomaticFacetGenerator({'desired-count': '1'}))
       .init();
     assertContainsAutomaticFacet();
   });
 
   it('should throw an error when areCollapsed property is invalid', () => {
     new TestFixture()
-      .with(addAutomaticFacetBuilder({'are-collapsed': 'potato'}))
+      .with(addAutomaticFacetGenerator({'are-collapsed': 'potato'}))
       .init();
     assertConsoleError();
     assertContainsComponentError(
       {
-        shadow: () => cy.get(automaticFacetBuilderComponent).shadow(),
+        shadow: () => cy.get(automaticFacetGeneratorComponent).shadow(),
       },
       true
     );
@@ -47,7 +49,7 @@ describe('Automatic Facet Builder Test Suites', () => {
   it('should collapse the facets when areCollapsed property is `true`', () => {
     new TestFixture()
       .with(
-        addAutomaticFacetBuilder({
+        addAutomaticFacetGenerator({
           'are-collapsed': 'true',
           'desired-count': '1',
         })
@@ -59,12 +61,36 @@ describe('Automatic Facet Builder Test Suites', () => {
   it('should not collapse the facets when areCollapsed property is `false`', () => {
     new TestFixture()
       .with(
-        addAutomaticFacetBuilder({
+        addAutomaticFacetGenerator({
           'are-collapsed': 'false',
           'desired-count': '1',
         })
       )
       .init();
     assertCollapseAutomaticFacets(false);
+  });
+
+  it('should display placeholders when no search has yet been executed', () => {
+    new TestFixture()
+      .with(
+        addAutomaticFacetGenerator({
+          'desired-count': '1',
+        })
+      )
+      .withoutFirstAutomaticSearch()
+      .init();
+    assertDisplayPlaceholder();
+  });
+
+  it('should display nothing when response is empty', () => {
+    new TestFixture()
+      .with(
+        addAutomaticFacetGenerator({
+          'desired-count': '1',
+        })
+      )
+      .withoutAutomaticFacets()
+      .init();
+    assertDisplayNothing();
   });
 });