+++ packages/atomic/src/components/search/smart-snippets/atomic-smart-snippet/atomic-smart-snippet.tsx
@@ -10,6 +10,7 @@ import {
   BindStateToController,
 } from '../../../../utils/initialization-utils';
 import {randomID} from '../../../../utils/utils';
+import {getAttributesFromLinkSlot} from '../../../common/result-link/attributes-slot';
 import {SmartSnippetCommon} from '../../../common/smart-snippets/atomic-smart-snippet/smart-snippet-common';
 import {Bindings} from '../../atomic-search-interface/atomic-search-interface';
 
@@ -28,6 +29,9 @@ import {Bindings} from '../../atomic-search-interface/atomic-search-interface';
  *   </template>
  * </atomic-smart-snippet>
  * ```
+ *
+ * @slot source-anchor-attributes - Lets you pass [attributes](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/a#attributes) down to anchor elements, overriding other attributes.
+ * To be used exclusively in anchor elements, such as: `<a slot="source-anchor-attributes" target="_blank"></a>`.
  *
  * @part smart-snippet - The wrapper of the entire smart snippet.
  * @part question - The header displaying the question that is answered by the found document excerpt.
@@ -102,6 +106,8 @@ export class AtomicSmartSnippet implements InitializableComponent {
     this.smartSnippetCommon = new SmartSnippetCommon({
       id: this.id,
       modalTagName: 'atomic-smart-snippet-feedback-modal',
+      getSourceAnchorAttributes: () =>
+        getAttributesFromLinkSlot(this.host, 'source-anchor-attributes'),
       getHost: () => this.host,
       getBindings: () => this.bindings,
       getModalRef: () => this.modalRef,