+++ packages/quantic/force-app/examples/main/lwc/exampleQuanticGeneratedAnswer/exampleQuanticGeneratedAnswer.js
@@ -1,0 +1,18 @@
+import {LightningElement, api, track} from 'lwc';
+
+export default class ExampleQuanticGeneratedAnswer extends LightningElement {
+  @api engineId = 'quantic-generated-answer-engine';
+  @track config = {};
+  isConfigured = false;
+
+  pageTitle = 'Quantic Generated Answer';
+  pageDescription =
+    'The QuanticGeneratedAnswer component automatically generates an answer using Coveo machine learning models to answer the query executed by the user.';
+  options = [
+  ];
+
+  handleTryItNow(evt) {
+    this.config = evt.detail;
+    this.isConfigured = true;
+  }
+}